+++ examples/tutorial_apm_simple/client/_api_client.py
@@ -18,8 +18,8 @@ class ApmSimpleClient:
     ApmSimpleClient
 
     Generated with:
-        pygen = 0.32.4
-        cognite-sdk = 7.5.1
+        pygen = 0.32.5
+        cognite-sdk = 7.8.3
         pydantic = 2.5.2
 
     Data Model:
@@ -36,7 +36,7 @@ class ApmSimpleClient:
         else:
             raise ValueError(f"Expected CogniteClient or ClientConfig, got {type(config_or_client)}")
         # The client name is used for aggregated logging of Pygen Usage
-        client.config.client_name = "CognitePygen:0.32.4"
+        client.config.client_name = "CognitePygen:0.32.5"
 
         view_by_write_class = {
             data_classes.AssetApply: dm.ViewId("tutorial_apm_simple", "Asset", "beb2bebdcbb4ad"),